pyproject.toml
[project]
name = "elt_core"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12,<3.13"
dependencies = [
        "dagster-cloud==1.11.2",
        "dagster-postgres==0.27.2",
        "dagster-k8s==0.27.2",
        "dagster-snowflake==0.27.2",
        "dagster-dbt==0.27.2",
        "dagster-dlt==0.27.2",
        "dagster-sling==0.27.2",
        "dbt-snowflake==1.10.0",
        "psutil",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[tool.dagster]
module_name = "elt_core.definitions"
code_location_name = "elt_core"

[tool.dg]
directory_type = "project"
project.root_module = "elt_core"


[dependency-groups]
dev = [
    "dagster-dg-cli>=1.11.2",
    "dagster-webserver>=1.11.2",
]
